refactor: created actions files for our story, our people and gallery pages
