+++ src/actions/galleryActions.ts
