+++ src/actions/ourStoryActions.ts
@@ -1,0 +1,1 @@
+/* If needed, place actions/functions related to the Our Story page here */
